Remove duplicate entries to match in game values.
Duplicate entries are caused by file name mismatch between base Mine and Slash and CTE2 operloader config. One is named "phys" and the other is "physical".

The values likely intended in CTE2 is physical_resist_big=5 and physical_resist=2 but due to the duplicate entry only physical_resist_big is being overwritten properly in the pack. The values seen in game reflect this as physical_resist_big=5 and physical_resist=4. This commit should make the planner match that.